# openapi.yaml
openapi: 3.0.3
info:
  title: NSIM Payment Network API
  description: |
    Payment network API for routing payment requests between merchants (SSIM) and card issuers (BSIM).

    NSIM supports **multi-bank routing** - cards from different BSIM instances (banks) are automatically
    routed to the correct issuer based on the card token format.

    ## Authentication
    All payment endpoints require a valid API key passed in the `X-API-Key` header.

    ## Multi-Bank Token Formats
    Card tokens contain the issuing bank identifier:
    - `wsim_bsim_xxx` → Routes to Bank Simulator (default BSIM)
    - `wsim_newbank_xxx` → Routes to NewBank BSIM instance
    - `ctok_xxx` → Consent token from default BSIM (OAuth flow)

    ## Payment Flow
    1. User completes OAuth consent at BSIM, selecting a card for payment
    2. BSIM issues a `cardToken` to the merchant (or WSIM issues wallet token)
    3. Merchant calls NSIM `/authorize` with the card token
    4. NSIM extracts the bank ID from the token and routes to the correct BSIM
    5. BSIM creates an authorization hold on the card
    6. Merchant captures the payment when goods are shipped
    7. Optional: void or refund as needed
  version: 1.3.0
  contact:
    name: BSIM Team
    url: https://github.com/jordancrombie/nsim

servers:
  - url: https://payment.banksim.ca/api/v1
    description: Production
  - url: http://localhost:3006/api/v1
    description: Local development

security:
  - ApiKeyAuth: []

paths:
  /health:
    get:
      summary: Health check
      description: Returns service health status
      tags:
        - Health
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  service:
                    type: string
                    example: nsim
                  timestamp:
                    type: string
                    format: date-time

  /payments/authorize:
    post:
      summary: Authorize a payment
      description: |
        Request authorization for a payment. Creates a hold on the card's available credit.
        Authorization expires after 7 days if not captured.
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizationRequest'
      responses:
        '200':
          description: Authorization approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationResponse'
        '400':
          description: Authorization declined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /payments/{transactionId}/capture:
    post:
      summary: Capture an authorized payment
      description: |
        Capture all or part of an authorized payment. The capture amount cannot exceed
        the authorized amount. Partial capture will release the remaining authorization.
      tags:
        - Payments
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
      responses:
        '200':
          description: Capture successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
        '400':
          description: Capture failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'

  /payments/{transactionId}/void:
    post:
      summary: Void an authorization
      description: |
        Void a pending authorization. This releases the held credit back to the cardholder.
        Cannot void a captured or already voided authorization.
      tags:
        - Payments
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoidRequest'
      responses:
        '200':
          description: Void successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoidResponse'
        '400':
          description: Void failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoidResponse'

  /payments/{transactionId}/refund:
    post:
      summary: Refund a captured payment
      description: |
        Refund all or part of a captured payment. Multiple partial refunds are supported
        up to the captured amount.
      tags:
        - Payments
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Refund successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'
        '400':
          description: Refund failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'

  /payments/{transactionId}:
    get:
      summary: Get transaction status
      description: Retrieve the current status and details of a transaction
      tags:
        - Payments
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /webhooks:
    post:
      summary: Register a webhook
      description: |
        Register a webhook endpoint to receive payment event notifications.
        Returns a secret for HMAC signature verification.
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRegistration'
      responses:
        '201':
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookConfig'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /webhooks/merchant/{merchantId}:
    get:
      summary: List merchant webhooks
      description: Get all webhook configurations for a merchant
      tags:
        - Webhooks
      parameters:
        - name: merchantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookConfig'

  /webhooks/{webhookId}:
    patch:
      summary: Update a webhook
      description: Update webhook URL, events, or active status
      tags:
        - Webhooks
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookUpdate'
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookConfig'
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a webhook
      description: Remove a webhook registration
      tags:
        - Webhooks
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Webhook deleted
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /diagnostics/analyze-token:
    post:
      summary: Analyze a card token
      description: |
        Debug utility to analyze a card token without processing a payment.
        Returns token type, format, and any decoded claims.
      tags:
        - Diagnostics
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cardToken
              properties:
                cardToken:
                  type: string
                  description: The token to analyze
                  example: wsim_newbank_abc123
      responses:
        '200':
          description: Token analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenAnalysis'

  /diagnostics/token-types:
    get:
      summary: List supported token types
      description: Returns documentation of all supported card token formats
      tags:
        - Diagnostics
      security: []
      responses:
        '200':
          description: Token type documentation
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokenTypes:
                    type: array
                    items:
                      type: object
                      properties:
                        prefix:
                          type: string
                        description:
                          type: string
                        example:
                          type: string

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    AuthorizationRequest:
      type: object
      required:
        - merchantId
        - amount
        - cardToken
        - orderId
      properties:
        merchantId:
          type: string
          description: Unique identifier for the merchant
          example: merchant_abc123
        merchantName:
          type: string
          description: Display name of the merchant
          example: SSIM Store
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Amount to authorize in the transaction currency
          example: 99.99
        currency:
          type: string
          default: CAD
          description: ISO 4217 currency code
          example: CAD
        cardToken:
          type: string
          description: |
            Token representing the consented card. Format determines routing:
            - `wsim_{bsimId}_xxx` - Wallet token from WSIM (routes to specified bank)
            - `ctok_xxx` - Consent token from OAuth flow (routes to default bank)
            - JWT format - May contain bsimId claim for routing
          example: wsim_bsim_abc123def456
        orderId:
          type: string
          description: Merchant's order identifier
          example: order_12345
        description:
          type: string
          description: Optional description of the purchase
          example: Widget purchase
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Optional key-value metadata
        agentContext:
          $ref: '#/components/schemas/AgentContext'

    AgentContext:
      type: object
      description: |
        SACP (SimToolBox Agent Commerce Protocol) context for AI agent-initiated transactions.
        Required fields when an AI agent initiates a payment on behalf of a human owner.
      required:
        - agentId
        - ownerId
        - humanPresent
      properties:
        agentId:
          type: string
          description: Unique agent identifier
          example: agent_abc123
        ownerId:
          type: string
          description: Human owner of the agent (WSIM user ID)
          example: user_xyz789
        humanPresent:
          type: boolean
          description: Was a human present when this transaction was initiated?
          example: false
        mandateId:
          type: string
          description: Reference to authorization mandate (optional)
          example: mandate_xyz789
        mandateType:
          type: string
          enum:
            - cart
            - intent
            - none
          description: Type of mandate authorizing this transaction
          example: cart

    AuthorizationResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
          description: NSIM transaction identifier
        status:
          $ref: '#/components/schemas/PaymentStatus'
        authorizationCode:
          type: string
          description: Authorization code from issuer (if approved)
          example: AUTH-ABC12345
        declineReason:
          type: string
          description: Reason for decline (if declined)
          example: Insufficient credit
        timestamp:
          type: string
          format: date-time

    CaptureRequest:
      type: object
      properties:
        amount:
          type: number
          format: decimal
          description: Amount to capture (defaults to full authorization)
          example: 99.99

    CaptureResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PaymentStatus'
        capturedAmount:
          type: number
          format: decimal
        timestamp:
          type: string
          format: date-time

    VoidRequest:
      type: object
      properties:
        reason:
          type: string
          description: Optional reason for voiding
          example: Customer cancelled order

    VoidResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PaymentStatus'
        timestamp:
          type: string
          format: date-time

    RefundRequest:
      type: object
      properties:
        amount:
          type: number
          format: decimal
          description: Amount to refund (defaults to remaining balance)
          example: 50.00
        reason:
          type: string
          description: Optional reason for refund
          example: Product returned

    RefundResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        refundId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PaymentStatus'
        refundedAmount:
          type: number
          format: decimal
        timestamp:
          type: string
          format: date-time

    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchantId:
          type: string
        merchantName:
          type: string
        orderId:
          type: string
        amount:
          type: number
          format: decimal
        currency:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
        authorizationCode:
          type: string
        capturedAmount:
          type: number
          format: decimal
        refundedAmount:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        # SACP: Agent context fields
        agentId:
          type: string
          description: Agent ID if transaction was initiated by an AI agent
        agentOwnerId:
          type: string
          description: Human owner of the agent (WSIM user ID)
        agentHumanPresent:
          type: boolean
          description: Was human present for this transaction?
        agentMandateId:
          type: string
          description: Reference to authorization mandate
        agentMandateType:
          type: string
          description: Type of mandate (cart, intent, none)

    PaymentStatus:
      type: string
      enum:
        - pending
        - authorized
        - captured
        - voided
        - refunded
        - declined
        - expired
        - failed

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message

    WebhookRegistration:
      type: object
      required:
        - merchantId
        - url
        - events
      properties:
        merchantId:
          type: string
          description: Merchant identifier
          example: ssim-regal-moose
        url:
          type: string
          format: uri
          description: HTTPS endpoint to receive webhook events
          example: https://store.example.com/webhooks/payment
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
          description: Events to subscribe to
          example: ['payment.authorized', 'payment.captured']
        secret:
          type: string
          description: Optional custom secret (auto-generated if not provided)

    WebhookConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchantId:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
        secret:
          type: string
          description: HMAC secret for signature verification
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    WebhookUpdate:
      type: object
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
        isActive:
          type: boolean

    WebhookEventType:
      type: string
      enum:
        - payment.authorized
        - payment.captured
        - payment.voided
        - payment.refunded
        - payment.declined
        - payment.expired
        - payment.failed

    TokenAnalysis:
      type: object
      properties:
        token:
          type: string
        format:
          type: string
          enum: [wsim_prefix, ctok_prefix, jwt, unknown]
        bsimId:
          type: string
          description: Detected bank ID for routing
        isJwt:
          type: boolean
        claims:
          type: object
          description: Decoded JWT claims (if applicable)
        warnings:
          type: array
          items:
            type: string
          description: Any issues detected with the token

tags:
  - name: Health
    description: Service health checks
  - name: Payments
    description: Payment processing operations
  - name: Webhooks
    description: Webhook registration and management
  - name: Diagnostics
    description: Token analysis and debugging utilities